feat(mesh): add desktop comment modal
